requirements.txt
flask
streamlit
rank_bm25
faiss-cpu
sentence-transformers
pandas
numpy
